[update] address pr comments
