+++ src/odemis/dataio/test/hdf5_test.py
@@ -30,7 +30,6 @@ from odemis.util import img
 import os
 import time
 import unittest
-from unittest.case import skip
 import json
 
 
@@ -1442,6 +1441,40 @@ class TestHDF5IO(unittest.TestCase):
         im = rthumbs[0]
         self.assertEqual(im.shape, tshape)
         self.assertEqual(im[0, 0].tolist(), [0, 255, 0])
+
+    def test_read_and_save_md_acq_recipes(self):
+        """
+        Checks that MD_ACQ_RECIPES metadata is saved and read back correctly.
+        """
+        size = (512, 256)
+        dtype = numpy.dtype("uint16")
+        metadata = {model.MD_SW_VERSION: "1.0-test",
+                    model.MD_HW_NAME: "fake hw",
+                    model.MD_DESCRIPTION: "test acq recipes",
+                    model.MD_ACQ_DATE: time.time(),
+                    model.MD_BPP: 12,
+                    model.MD_BINNING: (1, 1),
+                    model.MD_PIXEL_SIZE: (1e-6, 2e-5),
+                    model.MD_POS: (1e-3, -30e-3),
+                    model.MD_EXP_TIME: 1.2,
+                    model.MD_ACQ_RECIPES: "recipe1, recipe2",
+                    }
+
+        data = model.DataArray(numpy.zeros(size[::-1], dtype), metadata=metadata)
+
+        # export
+        hdf5.export(FILENAME, data)
+
+        # check it's here
+        st = os.stat(FILENAME)
+        self.assertGreater(st.st_size, 0)
+
+        # check data
+        rdata = hdf5.read_data(FILENAME)
+        self.assertEqual(len(rdata), 1)
+
+        im = rdata[0]
+        self.assertEqual(im.metadata[model.MD_ACQ_RECIPES], metadata[model.MD_ACQ_RECIPES])
 
 
 if __name__ == "__main__":
