+++ src/odemis/gui/cont/tabs/sparc_acquisition_tab.py
@@ -24,15 +24,9 @@ Odemis. If not, see http://www.gnu.org/licenses/.
 """
 
 import collections
-import json
 import logging
-import os
-import re
-from functools import partial
-from typing import Callable, Union, List, Dict
 
 import wx
-from wx.adv import EVT_WIZARD_BEFORE_PAGE_CHANGED, Wizard, WizardPageSimple
 
 import odemis.acq.stream as acqstream
 import odemis.gui.cont.acquisition as acqcont
@@ -41,7 +35,6 @@ import odemis.gui.model as guimod
 from odemis import model
 from odemis.acq import leech
 from odemis.acq.stream import (
-    NON_SETTINGS_VA,
     AngularSpectrumStream,
     ARSettingsStream,
     CLSettingsStream,
@@ -52,16 +45,13 @@ from odemis.acq.stream import (
     TemporalSpectrumStream,
 )
 from odemis.gui.comp import popup
-from odemis.gui.comp.foldpanelbar import EVT_CAPTIONBAR
-from odemis.gui.comp.settings import SettingsPanel
-from odemis.gui.comp.stream_bar import StreamBar
 from odemis.gui.comp.viewport import (
     MicroscopeViewport,
     PlotViewport,
     TemporalSpectrumViewport,
 )
 from odemis.gui.conf.data import get_stream_settings_config
-from odemis.gui.conf.file import CONF_PATH
+from odemis.gui.cont.recipes import SparcRecipesController
 from odemis.gui.cont.settings import (
     EBeamBlankerSettingsController,
     GunExciterSettingsController,
@@ -69,388 +59,7 @@ from odemis.gui.cont.settings import (
 from odemis.gui.cont.stream_bar import SparcStreamsController
 from odemis.gui.cont.tabs.tab import Tab
 from odemis.gui.model import TOOL_ACT_ZOOM_FIT, TOOL_NONE, TOOL_SPOT
-from odemis.gui.model.main_gui_data import MainGUIData
-from odemis.gui.model.tab_gui_data import SparcAcquisitionGUIData
-from odemis.gui.util import call_in_wx_main, get_home_folder
-
-TEMPLATES_DIR = os.path.join(get_home_folder(), ".config/odemis/templates")
-
-
-def show_validation_error(message: str):
-    """
-    Show a validation error message box.
-    :param message: The error message to display.
-    """
-    wx.MessageBox(message, "Validation Error", wx.ICON_ERROR)
-
-
-def show_error(message: str):
-    """
-    Show an error message box.
-    :param message: The error message to display.
-    """
-    wx.MessageBox(message, "Error", wx.ICON_ERROR | wx.OK)
-
-
-def show_warning(message: str):
-    """
-    Show a warning message box.
-    :param message: The warning message to display.
-    """
-    wx.MessageBox(message, "Warning", wx.ICON_WARNING | wx.OK)
-
-
-class FilesDialog(WizardPageSimple):
-    """
-    A wizard page to select one or multiple files, with optional checkbox to choose whether
-    to import/export the ROIs defined in the file(s) and remove current streams.
-
-    """
-    def __init__(self, parent, message: str, style: int):
-        """
-        :param parent: The parent window of the dialog.
-        :param message: The message to display at the top of the dialog.
-        :param style: The style of the file dialog (wx.FD_OPEN or wx.FD_SAVE, with optional wx.FD_MULTIPLE).
-        """
-        super().__init__(parent)
-        self.file_paths = None
-        self._message = message
-        self._style = style
-
-        sizer = wx.BoxSizer(wx.VERTICAL)
-
-        # Message at the top
-        lbl = wx.StaticText(self, label=self._message)
-        sizer.Add(lbl, 0, wx.ALL, 5)
-
-        # Browse button + file display in a horizontal row
-        row_sizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.file_path_ctrl = wx.ListBox(self, style=wx.LB_SINGLE)
-        row_sizer.Add(self.file_path_ctrl, 1, wx.ALIGN_CENTER_VERTICAL | wx.RIGHT, 5)
-
-        self.file_btn = wx.Button(self, label="Browse...")
-        self.file_btn.Bind(wx.EVT_BUTTON, self._on_browse)
-        row_sizer.Add(self.file_btn, 0, wx.ALIGN_CENTER_VERTICAL)
-        sizer.Add(row_sizer, 0, wx.EXPAND | wx.ALL, 5)
-
-        # ROI settings checkbox, to choose whether to import or export the ROIs defined
-        self.roi_cbox = wx.CheckBox(self, label="ROI settings")
-        self.roi_cbox.SetValue(True)
-        sizer.Add(self.roi_cbox, 0, wx.ALL, 5)
-
-        # Remove current streams checkbox, hidden by default
-        self.remove_streams_cbox = wx.CheckBox(self, label="Remove current streams")
-        self.remove_streams_cbox.SetValue(True)
-        self.remove_streams_cbox.Hide()
-        sizer.Add(self.remove_streams_cbox, 0, wx.ALL, 5)
-
-        self.SetSizerAndFit(sizer)
-
-    def _on_browse(self, evt):
-        """Open the file dialog to select one or multiple files, and display the selected file(s) in the list box."""
-        self.file_paths = self._get_filepaths_from_filedialog()
-        if self.file_paths is not None:
-            self.file_path_ctrl.Clear()
-            if self.file_paths:
-                for fp in self.file_paths:
-                    self.file_path_ctrl.Append(os.path.basename(fp))
-                self.file_path_ctrl.SetToolTip("\n".join(self.file_paths))
-            else:
-                self.file_path_ctrl.Append("No file selected")
-                self.file_path_ctrl.SetToolTip("")
-
-    def _get_filepaths_from_filedialog(self) -> Union[list[str], None]:
-        """
-        Open the file dialog to select one or multiple files, and return the selected file paths.
-        :return: A list of selected file paths, or None if the dialog was cancelled.
-        """
-        with wx.FileDialog(
-            parent=self,
-            message=self._message,
-            wildcard="JSON files (*.json)|*.json",
-            style=self._style,
-            defaultDir=CONF_PATH,
-            defaultFile="recipe.json",
-        ) as filedialog:
-            if filedialog.ShowModal() == wx.ID_CANCEL:
-                return None
-            if self._style & wx.FD_MULTIPLE:
-                paths = filedialog.GetPaths()
-            else:
-                paths = [filedialog.GetPath()]
-            return [p + ".json" if not p.lower().endswith(".json") else p for p in paths]
-
-    def validate(self) -> bool:
-        """
-        Validate the selected file paths.
-        Checks that at least one file is selected, and if in wx.FD_OPEN mode, that all selected files exist.
-        :return: True if the selected file paths are valid, False otherwise.
-        """
-        if not self.file_paths:
-            show_validation_error("Please select a .json file.")
-            return False
-        if self._style & wx.FD_OPEN:
-            if not all(os.path.exists(fp) for fp in self.file_paths):
-                show_validation_error("One or more selected files do not exist.")
-                return False
-        return True
-
-
-class StreamsPage(WizardPageSimple):
-    """
-    A wizard page to display stream settings entries, and let the user choose which stream(s) to import/export.
-    The stream settings entries are displayed in collapsible panels, one per stream, showing the settings of the
-    stream and letting the user modify them before importing/exporting.
-
-    """
-    def __init__(self, parent, main_data: MainGUIData, stctrl_menu_action: Dict[str, Callable]):
-        """
-        :param parent: The parent window of the page.
-        :param main_data: The main data model of the GUI, used to create the stream controllers for the
-            stream settings entries.
-        :param stctrl_menu_action: A dictionary mapping stream settings entry names to their corresponding
-            creation methods in the stream controller. Only the plugin custom stream settings entries with
-            a corresponding stream creation method will be displayed in the page.
-        """
-        super().__init__(parent)
-        self.snapshot_sse = []
-        self._stctrl_menu_action = stctrl_menu_action
-        tab_data = SparcAcquisitionGUIData(main_data)
-
-        # Create default SEMStream and AnchorDriftCorrector similar to the acquisition tab
-        tab_data.semStream = acqstream.SEMStream(
-            "Secondary electrons concurrent",
-            main_data.sed,
-            main_data.sed.data,
-            main_data.ebeam
-        )
-        tab_data.roa = tab_data.semStream.roi
-        tab_data.roa.value = acqstream.UNDEFINED_ROI
-        tab_data.driftCorrector = leech.AnchorDriftCorrector(tab_data.semStream.emitter,
-                                                             tab_data.semStream.detector)
-        tab_data.driftCorrector.roi.value = acqstream.UNDEFINED_ROI
-
-        # Custom scrolled window to be able to fit the stream settings entries panels
-        self._scrolled_win = wx.ScrolledWindow(self, style=wx.VSCROLL)
-        self._scrolled_win.SetScrollRate(0, 16)
-
-        stream_bar = StreamBar(self._scrolled_win, size=self.Parent.Size)
-        self.streambar_controller = SparcStreamsController(
-            tab_data,
-            stream_bar,
-            ignore_view=True
-        )
-        self.streambar_controller.add_action("Secondary electrons survey",
-                                             partial(self.streambar_controller._add_sem_stream,
-                                                     "Secondary electrons survey",
-                                                     self.streambar_controller._main_data_model.sed,
-                                                     add_to_view=True)
-                                             )
-        # Override the fit_streams method of the stream bar to handle the fitting of custom scrolled window
-        stream_bar.fit_streams = self._fit_streams
-
-        scrolled_sizer = wx.BoxSizer(wx.VERTICAL)
-        scrolled_sizer.Add(stream_bar, 1, wx.EXPAND)
-        self._scrolled_win.SetSizer(scrolled_sizer)
-
-        page_sizer = wx.BoxSizer(wx.VERTICAL)
-        page_sizer.Add(self._scrolled_win, 1, wx.EXPAND)
-        self.Bind(wx.EVT_SIZE, self.on_size)
-        self.SetSizerAndFit(page_sizer)
-
-    @call_in_wx_main
-    def _fit_streams(self):
-        if not self or self.IsBeingDeleted():
-            logging.debug("Not fitting streams because the streams page is being deleted")
-            return
-
-        stream_bar = self.streambar_controller._stream_bar
-        stream_bar._set_warning()
-        h = stream_bar.GetSizer().GetMinSize().GetHeight()
-        stream_bar.SetSize((-1, h))
-        self._scrolled_win.SetVirtualSize((-1, h))
-        stream_bar.Layout()
-        self._scrolled_win.Layout()
-
-    def on_size(self, evt):
-        self._fit_streams()
-        evt.Skip()
-
-    def _createBlSpectrumStreamActions(self):
-        """
-        Create the stream bar menu action for the "Spectrum" stream settings entries, depending on the
-        spectrometers available in the system.
-        """
-        main_data = self.streambar_controller._main_data_model
-        if main_data.spectrometers:
-            for sptm in main_data.spectrometers:
-                if len(main_data.spectrometers) <= 1:
-                    actname = "Spectrum"
-                else:
-                    actname = "Spectrum with %s" % (sptm.name,)
-
-                if actname in self._stctrl_menu_action:
-                    # Remove the standard action (which must have exactly the same name)
-                    self.streambar_controller.remove_action(actname)
-                    self.streambar_controller.add_action(actname, self._stctrl_menu_action[actname])
-
-    def _createLargeAreaSpectrumStreamActions(self):
-        """
-        Create the stream bar menu action for the "Large Area Spectrum" stream settings entries, depending on the
-        spectrometers available in the system and the lens switch configuration.
-        """
-        main_data = self.streambar_controller._main_data_model
-
-        if main_data.lens_switch and main_data.spectrometers:
-            for sptm in main_data.spectrometers:
-                # If the spectrometer is not affected by the lens switch, it means it's not on an optical path
-                # that supports large area FoV. It's probably on a second spectrograph. In this case,
-                # let's not give hope to the user, and don't provide this option.
-                if main_data.opm and not main_data.opm.affects(main_data.lens_switch.name, sptm.name):
-                    logging.info("Skipping LA spec %s as lens switch doesn't affect this detector", sptm.name)
-                    continue
-
-                if len(main_data.spectrometers) == 1:
-                    actname = "Large Area Spectrum"
-                else:
-                    actname = "Large Area Spectrum with %s" % (sptm.name,)
-
-                if actname in self._stctrl_menu_action:
-                    self.streambar_controller.add_action(actname, self._stctrl_menu_action[actname])
-
-    def _createSpectrumRawStreamAction(self):
-        """Create the stream bar menu action for the "Spectrum Raw" stream settings entry."""
-        actname = "Spectrum Raw"
-
-        if actname in self._stctrl_menu_action:
-            self.streambar_controller.add_action(actname, self._stctrl_menu_action[actname])
-
-    def _createSpectrumArbitraryOrderStreamAction(self):
-        """Create the stream bar menu action for the "Spectrum Arbitrary Order" stream settings entry."""
-        main_data = self.streambar_controller._main_data_model
-        if main_data.spectrometers:
-            for sptm in main_data.spectrometers:
-                if len(main_data.spectrometers) == 1:
-                    actname = "Spectrum Arbitrary Scan"
-                else:
-                    actname = "Spectrum Arbitrary Scan with %s" % (sptm.name,)
-
-                if actname in self._stctrl_menu_action:
-                    self.streambar_controller.add_action(actname, self._stctrl_menu_action[actname])
-
-    def _createCLIntensityCCDStreamAction(self):
-        """Create the stream bar menu action for the "CL intensity on CCD" stream settings entry."""
-        main_data = self.streambar_controller._main_data_model
-        if main_data.ccd:
-            actname = "CL intensity on CCD"
-
-            if actname in self._stctrl_menu_action:
-                self.streambar_controller.add_action(actname, self._stctrl_menu_action[actname])
-
-    def set_streams_settings_entries(self, streams_settings_entries: List[dict]):
-        """
-        Fill the stream bar with the stream settings entries.
-        :param streams_settings_entries: List of stream settings entries to fill the stream bar with.
-        """
-        if not isinstance(streams_settings_entries, list):
-            raise ValueError(f"Expected a list of stream settings entries, got {type(streams_settings_entries)}")
-        if not all(isinstance(sse, dict) for sse in streams_settings_entries):
-            raise ValueError("Expected a list of dicts as stream settings entries")
-
-        failed_streams = []
-        menu_actions = self.streambar_controller.menu_actions
-        tab_data = self.streambar_controller._tab_data_model
-        existing_streams_by_name = {s.name.value: s for s in tab_data.streams.value}
-
-        # Create the stream bar menu actions for custom plugin streams if they are not already created
-        # standard streams actions are created by default in the stream bar controller
-        for sse in streams_settings_entries:
-            name = sse["name"]
-            action_name = re.sub(r" \d+$", "", name)
-
-            if action_name not in menu_actions:
-                class_name = sse["class"]
-                if "BlSpectrumSettingsStream" == class_name:
-                    self._createBlSpectrumStreamActions()
-                elif "LASpectrumSettingsStream" == class_name:
-                    self._createLargeAreaSpectrumStreamActions()
-                elif "SpectrumRawSettingsStream" == class_name:
-                    self._createSpectrumRawStreamAction()
-                elif "SpectrumArbitraryOrderSettingsStream" == class_name:
-                    self._createSpectrumArbitraryOrderStreamAction()
-                elif "CL intensity on CCD" in name:
-                    self._createCLIntensityCCDStreamAction()
-
-        # For each stream settings entry, try to find an existing stream with the same name and update it,
-        # otherwise create a new stream with the corresponding stream creation method in the stream controller
-        for sse in streams_settings_entries:
-            name = sse["name"]
-            action_name = re.sub(r" \d+$", "", name)
-
-            if name == "Secondary electrons concurrent":
-                tab_data.semStream.set_settings_entries(sse)
-                continue
-            elif name == "Anchor drift corrector":
-                tab_data.driftCorrector.set_settings_entries(sse)
-                continue
-
-            # Update existing stream if one matches by name
-            if name in existing_streams_by_name:
-                existing_streams_by_name[name].set_settings_entries(sse)
-                continue
-
-            # Create a new stream
-            if action_name in menu_actions:
-                callback = menu_actions[action_name]
-                # Handle the "Secondary electrons survey" stream separately, as it doesn't follow the standard
-                # stream creation method signature (it cannot take the tab data and stream controller as arguments)
-                if action_name == "Secondary electrons survey":
-                    stream_ctrl = callback(settings_entries=sse)
-                else:
-                    # Handle both standard and custom plugin stream controller creation in the same way
-                    stream_ctrl = callback(
-                        tab_data=tab_data,
-                        stctrl=self.streambar_controller,
-                        settings_entries=sse
-                    )
-                stream_ctrl.stream_panel.show_visible_btn(False)
-                stream_ctrl.stream_panel.show_updated_btn(False)
-                stream_ctrl.stream_panel.collapse(True)
-            else:
-                if name == "Spot":
-                    continue
-                logging.debug(f"Stream settings entry with name '{name}' does not have a creation method. Skipping.")
-                failed_streams.append(name)
-
-        if failed_streams:
-            show_warning(f"The following streams could not be created:\n{', '.join(failed_streams)}")
-        self._fit_streams()
-
-    def take_snapshot(self) -> List[dict]:
-        """
-        Take a snapshot of the current stream settings entries in the stream bar, to be able to compare
-        with a future snapshot (eg, after importing a config) to know if the settings have been modified.
-        """
-        self.snapshot_sse = self.get_streams_settings_entries()
-
-    def validate(self) -> bool:
-        """Validate if there is at least one live stream available."""
-        ss = self.streambar_controller._tab_data_model.streams.value
-        if not ss:
-            show_validation_error("No streams available.")
-            return False
-        return True
-
-    def get_streams_settings_entries(self) -> List[dict]:
-        """
-        Get the current stream settings entries of the streams in the stream bar, to be able to save them
-        in a config file or compare with a previous snapshot.
-        """
-        ss = self.streambar_controller._tab_data_model.streams.value
-        sse = [s.get_settings_entries() for s in ss]
-        sse.append(self.streambar_controller._tab_data_model.semStream.get_settings_entries())
-        sse.append(self.streambar_controller._tab_data_model.driftCorrector.get_settings_entries())
-        return sse
+from odemis.gui.util import call_in_wx_main
 
 
 class SparcAcquisitionTab(Tab):
@@ -636,17 +245,6 @@ class SparcAcquisitionTab(Tab):
         )
         tab_data.driftCorrector.roi.subscribe(self._on_dc_roi, init=True)
 
-        # Acquisition recipes
-        self._prev_streams = set()
-        recipes_settings = SettingsPanel(panel.pnl_acq_recipes, size=(400, 90))
-        _, self.current_recipes = recipes_settings.add_readonly_field("Current recipes", "")
-        _, self.import_btn = recipes_settings.add_btn(label_text="Read from file(s)", btn_label="Import...")
-        _, self.export_btn = recipes_settings.add_btn(label_text="Write to file", btn_label="Export...")
-        self.export_btn.Bind(wx.EVT_BUTTON, self._export_streams_settings)
-        self.import_btn.Bind(wx.EVT_BUTTON, self._import_streams_settings)
-        cbar = wx.FindWindowByName("cbar_acq_recipes", panel)
-        cbar.Bind(EVT_CAPTIONBAR, self._on_recipes_captionbar)
-
         if main_data.pcd:
             # Create a "leech" that we can add/remove to the SEM stream
             self._pcd_acquirer = leech.ProbeCurrentAcquirer(main_data.pcd, main_data.pcd_sel)
@@ -713,6 +311,15 @@ class SparcAcquisitionTab(Tab):
             self._stream_controller
         )
 
+        # Acquisition recipes
+        self._recipes_controller = SparcRecipesController(
+            tab_data,
+            panel,
+            self.sem_stream,
+            self._stream_controller,
+            self._acquisition_controller
+        )
+
         # Force SEM view fit to content when magnification is updated
         if not main_data.ebeamControlsMag:
             main_data.ebeam.magnification.subscribe(self._onSEMMag)
@@ -747,44 +354,6 @@ class SparcAcquisitionTab(Tab):
 
         # Switch the last view to the most fitting one
         self.tab_data_model.visible_views.value[3] = br_view
-
-    def _on_any_va(self, _):
-        """
-        Called whenever a VA of any stream is updated. Used to update the current recipes display,
-        as any change in the stream settings may change the recipe.
-        """
-        if self.current_recipes.GetValue():
-            self._update_current_recipes_label(is_modified=True)
-
-    def _on_streams(self, streams):
-        """
-        Called when the .streams is updated, to subscribe to the VAs of the streams and update current recipes display.
-        """
-        streams = set(streams)
-
-        # remove subscription for streams that were deleted
-        for s in (self._prev_streams - streams):
-            for va in self._acquisition_controller._get_settings_vas(s):
-                va.unsubscribe(self._on_any_va)
-
-        # add subscription for new streams
-        for s in (streams - self._prev_streams):
-            for va in self._acquisition_controller._get_settings_vas(s):
-                va.subscribe(self._on_any_va)
-
-        self._prev_streams = streams
-
-    def _on_recipes_captionbar(self, evt):
-        """Toggle visibility of the recipes panel when the caption bar is clicked."""
-        cb = evt.get_bar()
-        pnl = self.panel.pnl_acq_recipes
-        if evt.get_fold_status():  # currently expanded → collapse
-            cb.collapse()
-            pnl.Hide()
-        else:  # currently collapsed → expand
-            cb.expand()
-            pnl.Show()
-        pnl.Parent.Parent.Layout()  # Layout the outer panel so scr_win_right absorbs the space
 
     def on_tool_change(self, tool):
         """ Ensure spot position is always defined when using the spot """
@@ -940,193 +509,6 @@ class SparcAcquisitionTab(Tab):
         #   play the stream again to set them again.
         if self.IsShown():
             self._stream_controller.pauseStreams()
-
-    def _update_current_recipes_label(self, is_modified: bool = False):
-        """
-        Update the current acquisition recipes label.
-        If is_modified is True, it means the settings were modified since the last import/export,
-        so add a "(modified)" label to warn the user that the current settings do not correspond to the listed recipes.
-
-        :param is_modified: whether the settings were modified since the last import/export.
-
-        """
-        if not self.tab_data_model.recipes:
-            self.current_recipes.SetValue("")
-            return
-        label = ", ".join(sorted(self.tab_data_model.recipes))
-        if is_modified:
-            label += " (modified)"
-        self.current_recipes.SetValue(label)
-
-    def _export_streams_settings(self, evt):
-        """
-        Export the current stream settings entries to a .json file, using a wizard to let the user select which streams
-        to export and the file path.
-        """
-        def validate_wizard_page(evt):
-            """Validate the wizard page before changing to the next page."""
-            page = evt.GetPage()
-            if page and evt.GetDirection():
-                if not page.validate():
-                    # If validation fails, prevent page change
-                    evt.Veto()
-
-        # "roi" is a non-settings VA by default. For SPARC the roi VA of any stream is relevant and we want to be able
-        # to export it, so we temporarily remove "roi" from the NON_SETTINGS_VA list. The stream_page.roi_cbox allows
-        # the user to choose whether to keep the roi settings or not.
-        if "roi" in NON_SETTINGS_VA:
-            NON_SETTINGS_VA.remove("roi")
-        orig_sem_sse = self.sem_stream.get_settings_entries()
-        message = "Export acquisition recipe to a .json file"
-        # Pause streams to avoid issues
-        self.streambar_controller.pauseStreams()
-
-        wizard = Wizard(self.streambar_controller._stream_bar, title="Export")
-        files_page = FilesDialog(wizard, message, wx.FD_SAVE | wx.FD_OVERWRITE_PROMPT)
-        streams_page = StreamsPage(wizard, self.tab_data_model.main, self.streambar_controller.menu_actions)
-        sse = [s.get_settings_entries() for s in self.tab_data_model.streams.value]
-        sse.append(self.tab_data_model.semStream.get_settings_entries())
-        sse.append(self.tab_data_model.driftCorrector.get_settings_entries())
-        streams_page.set_streams_settings_entries(sse)
-        streams_page.take_snapshot()
-
-        wizard.FitToPage(files_page)
-        wizard.FitToPage(streams_page)
-        WizardPageSimple.Chain(streams_page, files_page)
-        wizard.Bind(EVT_WIZARD_BEFORE_PAGE_CHANGED, validate_wizard_page)
-
-        try:
-            self.tab_data_model.streams.unsubscribe(self._on_streams)
-            self.tab_data_model.streams.unsubscribe(self._on_any_va)
-            if wizard.RunWizard(streams_page):
-                if not files_page.roi_cbox.IsChecked():
-                    NON_SETTINGS_VA.append("roi")
-                streams_settings_entries = streams_page.get_streams_settings_entries()
-                # Get the first file path (there should be only one in save mode)
-                file_path = files_page.file_paths[0]
-                with open(file_path, "w") as json_file:
-                    json.dump(streams_settings_entries, json_file, indent=2)
-                self.tab_data_model.recipes = {os.path.splitext(os.path.basename(file_path))[0]}
-                self._update_current_recipes_label()
-            else:
-                logging.debug("Export stream settings wizard was cancelled.")
-        except Exception as e:
-            logging.exception("An error occurred while exporting the stream settings.")
-            show_error(f"An error occurred while exporting the stream settings:\n{str(e)}")
-        finally:
-            self.tab_data_model.streams.subscribe(self._on_streams, init=True)
-            self.tab_data_model.streams.subscribe(self._on_any_va)
-            self.sem_stream.set_settings_entries(orig_sem_sse)
-            if "roi" not in NON_SETTINGS_VA:
-                NON_SETTINGS_VA.append("roi")
-            wizard.Destroy()
-
-    def _import_streams_settings(self, evt):
-        """
-        Import stream settings entries from a .json file(s), using a wizard to let the user select the file path and
-        which streams to import.
-        """
-        def before_wizard_page_changed(evt):
-            """
-            Validate the wizard page before changing to the next page, and if the next page is the streams settings page,
-            load the stream settings entries from the selected file(s) and display them in the streams settings page, to
-            let the user choose which ones to import.
-            """
-            page = evt.GetPage()
-            if page and evt.GetDirection():
-                if not page.validate():
-                    # If validation fails, prevent page change
-                    evt.Veto()
-                elif page is files_page:
-                    for file_path in files_page.file_paths:
-                        try:
-                            with open(file_path, "r") as json_file:
-                                config = json.load(json_file)
-                            streams_page.set_streams_settings_entries(config)
-                        except Exception as e:
-                            show_error(f"An error occurred while reading the file {file_path}:\n{str(e)}")
-                            evt.Veto()
-                            return
-                    # Take a snapshot of the current settings entries, to later know if they were modified by the user
-                    streams_page.take_snapshot()
-
-        # "roi" is a non-settings VA by default. For SPARC the roi VA of any stream is relevant and we want to be able
-        # to export it, so we temporarily remove "roi" from the NON_SETTINGS_VA list. The stream_page.roi_cbox allows
-        # the user to choose whether to keep the roi settings or not.
-        if "roi" in NON_SETTINGS_VA:
-            NON_SETTINGS_VA.remove("roi")
-        orig_sem_sse = self.sem_stream.get_settings_entries()
-        message = "Import acquisition recipes from .json file(s)"
-        # Pause streams to avoid issues
-        self.streambar_controller.pauseStreams()
-
-        wizard = Wizard(self.streambar_controller._stream_bar, title="Import")
-        files_page = FilesDialog(wizard, message, wx.FD_OPEN | wx.FD_FILE_MUST_EXIST | wx.FD_MULTIPLE)
-        files_page.remove_streams_cbox.Show()
-        streams_page = StreamsPage(wizard, self.tab_data_model.main, self.streambar_controller.menu_actions)
-
-        wizard.FitToPage(files_page)
-        wizard.FitToPage(streams_page)
-        WizardPageSimple.Chain(files_page, streams_page)
-        wizard.Bind(EVT_WIZARD_BEFORE_PAGE_CHANGED, before_wizard_page_changed)
-
-        try:
-            is_sem_stream = False
-            self.tab_data_model.streams.unsubscribe(self._on_streams)
-            self.tab_data_model.streams.unsubscribe(self._on_any_va)
-            if wizard.RunWizard(files_page):
-                if files_page.remove_streams_cbox.IsChecked():
-                    self.tab_data_model.recipes.clear()
-                    sps = self.streambar_controller._stream_bar.stream_panels
-                    for stream in self.tab_data_model.streams.value.copy():
-                        # SpotSEMStream and SEMStream should not be removed
-                        if not isinstance(stream, (acqstream.SpotSEMStream, acqstream.SEMStream)):
-                            self.streambar_controller.removeStream(stream)
-                            pnl = next((sp for sp in sps if sp.stream == stream), None)
-                            if pnl:
-                                # FIXME: removeStreamPanel should call removeStream on panel destroy
-                                # but it does not seem to work properly hence directly call remove_stream_panel for now
-                                self.streambar_controller._stream_bar.remove_stream_panel(pnl)
-
-                if not files_page.roi_cbox.IsChecked():
-                    NON_SETTINGS_VA.append("roi")
-                streams_settings_entries = streams_page.get_streams_settings_entries()
-
-                for sse in streams_settings_entries:
-                    name = sse["name"]
-                    action_name = re.sub(r" \d+$", "", name)
-
-                    if name == "Secondary electrons concurrent":
-                        self.tab_data_model.semStream.set_settings_entries(sse)
-                    elif name == "Anchor drift corrector":
-                        self.tab_data_model.driftCorrector.set_settings_entries(sse)
-                    elif name == "Secondary electrons survey":
-                        is_sem_stream = True
-                        self.sem_stream.set_settings_entries(sse)
-                    elif action_name in self.streambar_controller.menu_actions:
-                        callback = self.streambar_controller.menu_actions[action_name]
-                        callback(settings_entries=sse)
-                    else:
-                        logging.debug(f"No creation method defined for stream {name}. Skipping.")
-
-                if not is_sem_stream:
-                    self.sem_stream.set_settings_entries(orig_sem_sse)
-
-                for fp in files_page.file_paths:
-                    self.tab_data_model.recipes.add(os.path.splitext(os.path.basename(fp))[0])
-                modified = streams_page.snapshot_sse != streams_settings_entries
-                self._update_current_recipes_label(is_modified=modified)
-            else:
-                logging.debug("Import region wizard was cancelled.")
-        except Exception as e:
-            logging.exception("An error occurred while importing the stream settings.")
-            show_error(f"An error occurred while importing the stream settings:\n{str(e)}")
-        finally:
-            self.tab_data_model.streams.subscribe(self._on_streams, init=True)
-            self.tab_data_model.streams.subscribe(self._on_any_va)
-            if "roi" not in NON_SETTINGS_VA:
-                NON_SETTINGS_VA.append("roi")
-            wizard.Destroy()
 
     def Show(self, show=True):
         assert (show != self.IsShown())  # we assume it's only called when changed
